Safari-611.1.10/LayoutTests/http/tests/security/contentSecurityPolicy/script-src-none-inline-event-expected.txt

CONSOLE MESSAGE: Refused to execute a script for an inline event handler because 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.


--------
Frame: '<!--frame1-->'
--------
PASS

--------
Frame: '<!--frame2-->'
--------
PASS
